Extract `determineExecutionPlan` algorithm to a separate class
So that it can be understood and tested apart from the other execution plan
responsibilities.
