+++ subprojects/core/src/main/java/org/gradle/execution/plan/DefaultExecutionPlan.java
@@ -16,33 +16,17 @@
 
 package org.gradle.execution.plan;
 
-import com.google.common.collect.HashMultimap;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Iterables;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import org.gradle.api.Action;
 import org.gradle.api.BuildCancelledException;
-import org.gradle.api.CircularReferenceException;
-import org.gradle.api.GradleException;
 import org.gradle.api.NonNullApi;
 import org.gradle.api.Task;
-import org.gradle.api.internal.TaskInternal;
-import org.gradle.api.internal.project.ProjectInternal;
-import org.gradle.api.internal.tasks.TaskDestroyablesInternal;
-import org.gradle.api.internal.tasks.TaskLocalStateInternal;
-import org.gradle.api.internal.tasks.properties.OutputFilePropertyType;
-import org.gradle.api.internal.tasks.properties.PropertyValue;
-import org.gradle.api.internal.tasks.properties.PropertyVisitor;
-import org.gradle.api.internal.tasks.properties.PropertyWalker;
 import org.gradle.api.specs.Spec;
 import org.gradle.api.specs.Specs;
 import org.gradle.internal.Pair;
-import org.gradle.internal.graph.CachingDirectedGraphWalker;
-import org.gradle.internal.graph.DirectedGraphRenderer;
-import org.gradle.internal.logging.text.StyledTextOutput;
-import org.gradle.internal.reflect.validation.TypeValidationContext;
 import org.gradle.internal.resources.ResourceLock;
 import org.gradle.internal.resources.ResourceLockCoordinationService;
 import org.gradle.internal.work.WorkerLeaseRegistry;
@@ -50,7 +34,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nullable;
-import java.io.StringWriter;
 import java.util.AbstractCollection;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
@@ -61,17 +44,14 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Consumer;
 import java.util.stream.Stream;
 
-import static com.google.common.collect.Lists.newLinkedList;
 import static com.google.common.collect.Sets.newIdentityHashSet;
 import static java.lang.String.format;
-import static org.gradle.execution.plan.NodeSets.sortedListOf;
 
 /**
  * The mutation methods on this implementation are NOT threadsafe, and callers must synchronize access to these methods.
@@ -239,91 +219,7 @@ public class DefaultExecutionPlan implements ExecutionPlan, WorkSource<Node> {
 
     @Override
     public void determineExecutionPlan() {
-        updateFinalizerGroups();
-
-        LinkedList<NodeInVisitingSegment> nodeQueue = newLinkedList();
-        int visitingSegmentCounter = 0;
-        for (Node node : entryNodes) {
-            nodeQueue.add(new NodeInVisitingSegment(node, visitingSegmentCounter++));
-        }
-
-        HashMultimap<Node, Integer> visitingNodes = HashMultimap.create();
-        Deque<GraphEdge> walkedShouldRunAfterEdges = new ArrayDeque<>();
-        Deque<Node> path = new ArrayDeque<>();
-        Map<Node, Integer> planBeforeVisiting = new HashMap<>();
-
-        while (!nodeQueue.isEmpty()) {
-            NodeInVisitingSegment nodeInVisitingSegment = nodeQueue.peekFirst();
-            int currentSegment = nodeInVisitingSegment.visitingSegment;
-            Node node = nodeInVisitingSegment.node;
-
-            if (node.isDoNotIncludeInPlan() || nodeMapping.contains(node)) {
-                // Discard the node because it has already been visited or should not be included, for example:
-                // - it has already executed in another execution plan
-                // - it is reachable only via a must-run-after or should-run-after edge
-                // - it is filtered
-                nodeQueue.removeFirst();
-                visitingNodes.remove(node, currentSegment);
-                maybeRemoveProcessedShouldRunAfterEdge(walkedShouldRunAfterEdges, node);
-                continue;
-            }
-
-            boolean alreadyVisited = visitingNodes.containsKey(node);
-            visitingNodes.put(node, currentSegment);
-
-            if (!alreadyVisited) {
-                // Have not seen this node before - add its dependencies to the head of the queue and leave this
-                // node in the queue
-                recordEdgeIfArrivedViaShouldRunAfter(walkedShouldRunAfterEdges, path, node);
-                removeShouldRunAfterSuccessorsIfTheyImposeACycle(visitingNodes, nodeInVisitingSegment);
-                takePlanSnapshotIfCanBeRestoredToCurrentTask(planBeforeVisiting, node);
-
-                // Add any finalizers to the queue just after the current node
-                for (Node finalizer : node.getFinalizers()) {
-                    addFinalizerToQueue(nodeQueue, visitingSegmentCounter++, finalizer);
-                }
-
-                for (Node successor : node.getAllSuccessorsInReverseOrder()) {
-                    if (visitingNodes.containsEntry(successor, currentSegment)) {
-                        if (!walkedShouldRunAfterEdges.isEmpty()) {
-                            //remove the last walked should run after edge and restore state from before walking it
-                            GraphEdge toBeRemoved = walkedShouldRunAfterEdges.pop();
-                            // Should run after edges only exist between tasks, so this cast is safe
-                            TaskNode sourceTask = (TaskNode) toBeRemoved.from;
-                            TaskNode targetTask = (TaskNode) toBeRemoved.to;
-                            sourceTask.removeShouldSuccessor(targetTask);
-                            restorePath(path, toBeRemoved);
-                            restoreQueue(nodeQueue, visitingNodes, toBeRemoved);
-                            restoreExecutionPlan(planBeforeVisiting, toBeRemoved);
-                            break;
-                        } else {
-                            onOrderingCycle(successor, node);
-                        }
-                    }
-                    nodeQueue.addFirst(new NodeInVisitingSegment(successor, currentSegment));
-                }
-                path.push(node);
-            } else {
-                // Have visited this node's dependencies - add it to the end of the plan
-                nodeQueue.removeFirst();
-                maybeRemoveProcessedShouldRunAfterEdge(walkedShouldRunAfterEdges, node);
-                visitingNodes.remove(node, currentSegment);
-                path.pop();
-                nodeMapping.add(node);
-
-                for (Node dependency : node.getDependencySuccessors()) {
-                    dependency.getMutationInfo().consumingNodes.add(node);
-                }
-            }
-        }
-
-        for (Node node : nodeMapping) {
-            node.maybeUpdateOrdinalGroup();
-            createOrdinalRelationships(node);
-        }
-
-        ordinalNodeAccess.createInterNodeRelationships();
-        nodeMapping.addAll(ordinalNodeAccess.getAllNodes());
+        new DetermineExecutionPlanAction(nodeMapping, ordinalNodeAccess, entryNodes, finalizers).run();
         dependencyResolver.clear();
         executionQueue.setNodes(nodeMapping);
     }
@@ -336,7 +232,6 @@ public class DefaultExecutionPlan implements ExecutionPlan, WorkSource<Node> {
             node.updateAllDependenciesComplete();
             maybeNodeReady(node);
         }
-
         lockCoordinator.addLockReleaseListener(resourceUnlockListener);
     }
 
@@ -344,23 +239,6 @@ public class DefaultExecutionPlan implements ExecutionPlan, WorkSource<Node> {
     public WorkSource<Node> asWorkSource() {
         // For now
         return this;
-    }
-
-    private void addFinalizerToQueue(LinkedList<NodeInVisitingSegment> nodeQueue, int visitingSegmentCounter, Node finalizer) {
-        int insertPosition = 1;
-        int pos = 0;
-        for (NodeInVisitingSegment segment : nodeQueue) {
-            if (segment.node == finalizer) {
-                // Already later in the queue
-                return;
-            }
-            // Need to insert the finalizer immediately after the last node that it finalizes
-            if (finalizer.getFinalizingSuccessors().contains(segment.node) && pos > insertPosition) {
-                insertPosition = pos;
-            }
-            pos++;
-        }
-        nodeQueue.add(insertPosition, new NodeInVisitingSegment(finalizer, visitingSegmentCounter));
     }
 
     @Override
@@ -387,170 +265,6 @@ public class DefaultExecutionPlan implements ExecutionPlan, WorkSource<Node> {
         if (!(resourceLock instanceof WorkerLeaseRegistry.WorkerLease) && maybeNodesReady) {
             maybeNodesSelectable = true;
         }
-    }
-
-    private void createOrdinalRelationships(Node node) {
-        if (!(node instanceof TaskNode)) {
-            return;
-        }
-
-        OrdinalGroup ordinal = node.getOrdinal();
-        if (ordinal == null) {
-            return;
-        }
-
-        TaskNode taskNode = (TaskNode) node;
-        TaskClassifier taskClassifier = classifyTask(taskNode);
-
-        if (taskClassifier.isDestroyer()) {
-            // Create (or get) a destroyer ordinal node that depends on the dependencies of this task node
-            OrdinalNode ordinalNode = ordinalNodeAccess.getOrCreateDestroyableLocationNode(ordinal);
-            ordinalNode.addDependenciesFrom(taskNode);
-
-            Node precedingProducersNode = ordinalNodeAccess.getPrecedingProducerLocationNode(ordinal);
-            if (precedingProducersNode != null) {
-                // Depend on any previous producer ordinal nodes (i.e. any producer ordinal nodes with a lower ordinal)
-                taskNode.addDependencySuccessor(precedingProducersNode);
-            }
-        } else if (taskClassifier.isProducer()) {
-            // Create (or get) a producer ordinal node that depends on the dependencies of this task node
-            OrdinalNode ordinalNode = ordinalNodeAccess.getOrCreateOutputLocationNode(ordinal);
-            ordinalNode.addDependenciesFrom(taskNode);
-
-            Node precedingDestroyersNode = ordinalNodeAccess.getPrecedingDestroyerLocationNode(ordinal);
-            if (precedingDestroyersNode != null) {
-                // Depend on any previous destroyer ordinal nodes (i.e. any destroyer ordinal nodes with a lower ordinal)
-                taskNode.addDependencySuccessor(precedingDestroyersNode);
-            }
-        }
-    }
-
-    /**
-     * Walk the properties of the task to determine if it is a destroyer or a producer (or neither).
-     */
-    private TaskClassifier classifyTask(TaskNode taskNode) {
-        TaskClassifier taskClassifier = new TaskClassifier();
-        TaskInternal task = taskNode.getTask();
-
-        PropertyWalker propertyWalker = propertyWalkerOf(task);
-        propertyWalker.visitProperties(task, TypeValidationContext.NOOP, taskClassifier);
-        task.getOutputs().visitRegisteredProperties(taskClassifier);
-        if (taskClassifier.isDestroyer()) {
-            // avoid walking further properties after discovering the task is destroyer
-            return taskClassifier;
-        }
-
-        ((TaskDestroyablesInternal) task.getDestroyables()).visitRegisteredProperties(taskClassifier);
-        if (taskClassifier.isDestroyer()) {
-            // avoid walking further properties after discovering the task is destroyer
-            return taskClassifier;
-        }
-
-        ((TaskLocalStateInternal) task.getLocalState()).visitRegisteredProperties(taskClassifier);
-
-        return taskClassifier;
-    }
-
-    private PropertyWalker propertyWalkerOf(TaskInternal task) {
-        return ((ProjectInternal) task.getProject()).getServices().get(PropertyWalker.class);
-    }
-
-    private void maybeRemoveProcessedShouldRunAfterEdge(Deque<GraphEdge> walkedShouldRunAfterEdges, Node node) {
-        GraphEdge edge = walkedShouldRunAfterEdges.peek();
-        if (edge != null && edge.to.equals(node)) {
-            walkedShouldRunAfterEdges.pop();
-        }
-    }
-
-    private void restoreExecutionPlan(Map<Node, Integer> planBeforeVisiting, GraphEdge toBeRemoved) {
-        int count = planBeforeVisiting.get(toBeRemoved.from);
-        nodeMapping.retainFirst(count);
-    }
-
-    private void restoreQueue(Deque<NodeInVisitingSegment> nodeQueue, HashMultimap<Node, Integer> visitingNodes, GraphEdge toBeRemoved) {
-        NodeInVisitingSegment nextInQueue = null;
-        while (nextInQueue == null || !toBeRemoved.from.equals(nextInQueue.node)) {
-            nextInQueue = nodeQueue.peekFirst();
-            visitingNodes.remove(nextInQueue.node, nextInQueue.visitingSegment);
-            if (!toBeRemoved.from.equals(nextInQueue.node)) {
-                nodeQueue.removeFirst();
-            }
-        }
-    }
-
-    private void restorePath(Deque<Node> path, GraphEdge toBeRemoved) {
-        Node removedFromPath = null;
-        while (!toBeRemoved.from.equals(removedFromPath)) {
-            removedFromPath = path.pop();
-        }
-    }
-
-    private void removeShouldRunAfterSuccessorsIfTheyImposeACycle(final HashMultimap<Node, Integer> visitingNodes, final NodeInVisitingSegment nodeWithVisitingSegment) {
-        Node node = nodeWithVisitingSegment.node;
-        if (!(node instanceof TaskNode)) {
-            return;
-        }
-        Iterables.removeIf(
-            ((TaskNode) node).getShouldSuccessors(),
-            input -> visitingNodes.containsEntry(input, nodeWithVisitingSegment.visitingSegment)
-        );
-    }
-
-    private void takePlanSnapshotIfCanBeRestoredToCurrentTask(Map<Node, Integer> planBeforeVisiting, Node node) {
-        if (node instanceof TaskNode && !((TaskNode) node).getShouldSuccessors().isEmpty()) {
-            planBeforeVisiting.put(node, nodeMapping.size());
-        }
-    }
-
-    private void recordEdgeIfArrivedViaShouldRunAfter(Deque<GraphEdge> walkedShouldRunAfterEdges, Deque<Node> path, Node node) {
-        if (!(node instanceof TaskNode)) {
-            return;
-        }
-        Node previous = path.peek();
-        if (previous instanceof TaskNode && ((TaskNode) previous).getShouldSuccessors().contains(node)) {
-            walkedShouldRunAfterEdges.push(new GraphEdge(previous, node));
-        }
-    }
-
-    private void onOrderingCycle(Node successor, Node currentNode) {
-        List<Set<Node>> cycles = findCycles(successor);
-        if (cycles.isEmpty()) {
-            // TODO: This isn't correct. This means that we've detected a cycle while determining the execution plan, but the graph walker did not find one.
-            // https://github.com/gradle/gradle/issues/2293
-            throw new GradleException("Misdetected cycle between " + currentNode + " and " + successor + ". Help us by reporting this to https://github.com/gradle/gradle/issues/2293");
-        }
-        StringWriter cycleString = renderOrderingCycle(cycles.get(0));
-        throw new CircularReferenceException(format("Circular dependency between the following tasks:%n%s", cycleString));
-    }
-
-    private List<Set<Node>> findCycles(Node successor) {
-        CachingDirectedGraphWalker<Node, Void> graphWalker = new CachingDirectedGraphWalker<>((node, values, connectedNodes) -> {
-            connectedNodes.addAll(node.getDependencySuccessors());
-            if (node instanceof TaskNode) {
-                TaskNode taskNode = (TaskNode) node;
-                connectedNodes.addAll(taskNode.getMustSuccessors());
-                connectedNodes.addAll(taskNode.getFinalizingSuccessors());
-            }
-        });
-        graphWalker.add(successor);
-        return graphWalker.findCycles();
-    }
-
-    private StringWriter renderOrderingCycle(Set<Node> nodes) {
-        List<Node> cycle = sortedListOf(nodes);
-
-        DirectedGraphRenderer<Node> graphRenderer = new DirectedGraphRenderer<>(
-            (it, output) -> output.withStyle(StyledTextOutput.Style.Identifier).text(it),
-            (it, values, connectedNodes) -> {
-                for (Node dependency : cycle) {
-                    if (it.hasHardSuccessor(dependency)) {
-                        connectedNodes.add(dependency);
-                    }
-                }
-            });
-        StringWriter writer = new StringWriter();
-        graphRenderer.renderTo(cycle.get(0), writer);
-        return writer;
     }
 
     @Override
@@ -992,36 +706,6 @@ public class DefaultExecutionPlan implements ExecutionPlan, WorkSource<Node> {
         }
     }
 
-    private void updateFinalizerGroups() {
-        // Collect the finalizers and their dependencies so that each node is ordered before all of its dependencies
-        LinkedList<Node> nodes = new LinkedList<>();
-        Set<Node> visiting = new HashSet<>();
-        Set<Node> visited = new HashSet<>();
-        Deque<Node> queue = new ArrayDeque<>(finalizers);
-        while (!queue.isEmpty()) {
-            Node node = queue.peek();
-            if (visited.contains(node)) {
-                // Already visited node, skip
-                queue.remove();
-            } else if (visiting.add(node)) {
-                // Haven't seen this node
-                for (Node successor : node.getDependencySuccessors()) {
-                    queue.addFirst(successor);
-                }
-            } else {
-                // Have visited the dependencies of this node, add it to the start of the list (so that it is earlier in the list that
-                // all of its dependencies)
-                visiting.remove(node);
-                visited.add(node);
-                nodes.addFirst(node);
-            }
-        }
-        for (Node node : nodes) {
-            node.updateGroupOfFinalizer();
-        }
-        finalizers.clear();
-    }
-
     private void handleFailure(Node node) {
         Throwable executionFailure = node.getExecutionFailure();
         if (executionFailure != null) {
@@ -1095,27 +779,7 @@ public class DefaultExecutionPlan implements ExecutionPlan, WorkSource<Node> {
         return nodeMapping.getNumberOfPublicNodes();
     }
 
-    private static class GraphEdge {
-        private final Node from;
-        private final Node to;
-
-        private GraphEdge(Node from, Node to) {
-            this.from = from;
-            this.to = to;
-        }
-    }
-
-    private static class NodeInVisitingSegment {
-        private final Node node;
-        private final int visitingSegment;
-
-        private NodeInVisitingSegment(Node node, int visitingSegment) {
-            this.node = node;
-            this.visitingSegment = visitingSegment;
-        }
-    }
-
-    private static class NodeMapping extends AbstractCollection<Node> {
+    static class NodeMapping extends AbstractCollection<Node> {
         private final Map<Task, LocalTaskNode> taskMapping = Maps.newLinkedHashMap();
         private final Set<Node> nodes = Sets.newLinkedHashSet();
 
@@ -1189,35 +853,7 @@ public class DefaultExecutionPlan implements ExecutionPlan, WorkSource<Node> {
         }
     }
 
-    private static class TaskClassifier extends PropertyVisitor.Adapter {
-        private boolean isProducer;
-        private boolean isDestroyer;
-
-        @Override
-        public void visitOutputFileProperty(String propertyName, boolean optional, PropertyValue value, OutputFilePropertyType filePropertyType) {
-            isProducer = true;
-        }
-
-        @Override
-        public void visitDestroyableProperty(Object value) {
-            isDestroyer = true;
-        }
-
-        @Override
-        public void visitLocalStateProperty(Object value) {
-            isProducer = true;
-        }
-
-        public boolean isProducer() {
-            return isProducer;
-        }
-
-        public boolean isDestroyer() {
-            return isDestroyer;
-        }
-    }
-
-    private static class ExecutionQueue {
+    static class ExecutionQueue {
         private final List<Node> nodes = new ArrayList<>();
         private int pos = 0;
